// src/shared/types.d.ts
interface oCommits {
  reponame: string;
  additions: number;
  deletions: number;
  total: number;
  message: string;
  time: string;
  committer: string;
  commit: string;
}